Fixed [SAIC-241]: removed temp option from config
Created temp dir for check bandwidth step.
Created temp dir for ssh chunks operations.
Removed temp option from config and workaround for map_compute_info.

+++ cloudferrylib/utils/files.py
@@ -68,9 +68,22 @@ class RemoteDir(object):
         return self
 
 
+class GetTempDir(object):
+    def __init__(self, runner, prefix):
+        self.runner = runner
+        self.prefix = prefix
+
+    def get(self):
+        cmd = "mktemp -udt %s_XXXX" % self.prefix
+        return self.runner.run(cmd)
+
 def _unlink(filename):
     return "rm -f {file}".format(file=filename)
 
 
 def _unlink_dir(dirname):
-    return "rm -rf {dir}".format(dir=dirname)
+    if len(dirname) > 1:
+        return "rm -rf {dir}".format(dir=dirname)
+    else:
+        raise RuntimeError('Wrong dirname %s, stopping' % dirname)
+    return
